+++ crates/rspack_plugin_circular_dependencies/src/lib.rs
@@ -1,272 +1,9 @@
-use rspack_collections::{IdentifierMap, IdentifierSet};
-use rspack_core::{
-  Compilation, CompilationOptimizeModules, ContextTypePrefix, DependencyType, ModuleGraph,
-  ModuleIdentifier, Plugin, incremental::Mutation,
-};
-use rspack_error::{Diagnostic, Result};
-use rspack_hook::{plugin, plugin_hook};
+mod circular_dependency_rspack_plugin;
+mod circular_modules_info_plugin;
 
-mod deprecated;
-
-pub use deprecated::{
+pub use circular_dependency_rspack_plugin::{
   CircularDependencyIgnoredConnection, CircularDependencyIgnoredConnectionEntry,
   CircularDependencyRspackPlugin, CircularDependencyRspackPluginOptions, CompilationHookFn,
   CycleHandlerFn,
 };
-
-struct CycleGraph {
-  modules: Vec<ModuleIdentifier>,
-  edges: Vec<Vec<usize>>,
-}
-
-impl CycleGraph {
-  fn build(module_graph: &ModuleGraph) -> Self {
-    let mut modules = module_graph
-      .modules()
-      .filter_map(|(&id, module)| module.source().is_some().then_some(id))
-      .collect::<Vec<_>>();
-    modules.sort_unstable();
-
-    let mut indexes = IdentifierMap::default();
-    indexes.reserve(modules.len());
-    for (index, id) in modules.iter().enumerate() {
-      indexes.insert(*id, index);
-    }
-
-    let mut edges = vec![Vec::new(); modules.len()];
-    for (index, module_id) in modules.iter().enumerate() {
-      let module_edges = &mut edges[index];
-      for connection in module_graph.get_outgoing_connections(module_id) {
-        let dependency = module_graph.dependency_by_id(&connection.dependency_id);
-        if should_ignore_dependency_type(*dependency.dependency_type()) {
-          continue;
-        }
-
-        let target_id = *connection.module_identifier();
-        if target_id == *module_id {
-          continue;
-        }
-
-        if let Some(&target_index) = indexes.get(&target_id) {
-          module_edges.push(target_index);
-        }
-      }
-
-      module_edges.sort_unstable();
-      module_edges.dedup();
-    }
-
-    Self { modules, edges }
-  }
-}
-
-#[derive(Clone, Copy, Default)]
-struct NodeState {
-  index: Option<usize>,
-  low_link: usize,
-  on_stack: bool,
-}
-
-struct ConnectFrame {
-  module_index: usize,
-  next_edge_index: usize,
-  parent_index: Option<usize>,
-}
-
-struct CycleDetector<'a> {
-  graph: &'a CycleGraph,
-  next_index: usize,
-  states: Vec<NodeState>,
-  stack: Vec<usize>,
-  circular_modules: IdentifierSet,
-}
-
-impl<'a> CycleDetector<'a> {
-  fn new(graph: &'a CycleGraph) -> Self {
-    let size = graph.modules.len();
-    Self {
-      graph,
-      next_index: 0,
-      states: vec![NodeState::default(); size],
-      stack: Vec::with_capacity(size),
-      circular_modules: IdentifierSet::default(),
-    }
-  }
-
-  fn find_circular_modules(mut self) -> IdentifierSet {
-    for module_index in 0..self.graph.modules.len() {
-      if self.states[module_index].index.is_none() {
-        self.connect(module_index);
-      }
-    }
-
-    self.circular_modules
-  }
-
-  fn connect(&mut self, module_index: usize) {
-    self.start_connect(module_index);
-    let mut visit_stack = vec![ConnectFrame {
-      module_index,
-      next_edge_index: 0,
-      parent_index: None,
-    }];
-
-    while !visit_stack.is_empty() {
-      let next_target = {
-        let frame = visit_stack
-          .last_mut()
-          .expect("visit stack should not be empty");
-        let module_index = frame.module_index;
-        if frame.next_edge_index < self.graph.edges[module_index].len() {
-          let target_index = self.graph.edges[module_index][frame.next_edge_index];
-          frame.next_edge_index += 1;
-          Some((module_index, target_index))
-        } else {
-          None
-        }
-      };
-
-      if let Some((module_index, target_index)) = next_target {
-        if self.states[target_index].index.is_none() {
-          self.start_connect(target_index);
-          visit_stack.push(ConnectFrame {
-            module_index: target_index,
-            next_edge_index: 0,
-            parent_index: Some(module_index),
-          });
-        } else if self.states[target_index].on_stack {
-          self.states[module_index].low_link = self.states[module_index]
-            .low_link
-            .min(self.states[target_index].index.expect("indexed"));
-        }
-        continue;
-      }
-
-      let frame = visit_stack.pop().expect("visit stack should not be empty");
-      let module_index = frame.module_index;
-      self.finish_connect(module_index);
-
-      if let Some(parent_index) = frame.parent_index {
-        self.states[parent_index].low_link = self.states[parent_index]
-          .low_link
-          .min(self.states[module_index].low_link);
-      }
-    }
-  }
-
-  fn start_connect(&mut self, module_index: usize) {
-    let index = self.next_index;
-    self.next_index += 1;
-    self.states[module_index] = NodeState {
-      index: Some(index),
-      low_link: index,
-      on_stack: true,
-    };
-    self.stack.push(module_index);
-  }
-
-  fn finish_connect(&mut self, module_index: usize) {
-    if self.states[module_index].low_link == self.states[module_index].index.expect("indexed") {
-      let mut component = Vec::new();
-      loop {
-        let current = self.stack.pop().expect("root should be on the stack");
-        self.states[current].on_stack = false;
-        component.push(current);
-        if current == module_index {
-          break;
-        }
-      }
-
-      if component.len() > 1 {
-        self
-          .circular_modules
-          .extend(component.into_iter().map(|index| self.graph.modules[index]));
-      }
-    }
-  }
-}
-
-#[plugin]
-#[derive(Debug)]
-pub struct CircularModulesInfoPlugin {}
-
-impl CircularModulesInfoPlugin {
-  pub fn new() -> Self {
-    Self::new_inner()
-  }
-}
-
-fn should_ignore_dependency_type(ty: DependencyType) -> bool {
-  matches!(
-    ty,
-    // Async boundaries. These edges do not synchronously evaluate the target
-    // module while the current module is initializing.
-    DependencyType::DynamicImport
-      | DependencyType::DynamicImportEager
-      | DependencyType::DynamicImportWeak
-      | DependencyType::ImportContext
-      | DependencyType::LazyImport
-      | DependencyType::ContextElement(ContextTypePrefix::Import)
-      | DependencyType::RequireEnsure
-      | DependencyType::RequireEnsureItem
-      // Resolve/id-only references. They need a target module for resolution or
-      // ID generation, but do not evaluate the target module.
-      | DependencyType::ImportMetaResolve
-      | DependencyType::ImportMetaResolveContext
-      | DependencyType::RequireResolve
-      | DependencyType::RequireResolveContext
-      | DependencyType::IsIncluded
-      // HMR accept/decline references are invoked by the hot runtime later, not
-      // by normal module initialization.
-      | DependencyType::ImportMetaHotAccept
-      | DependencyType::ImportMetaHotDecline
-      | DependencyType::ModuleHotAccept
-      | DependencyType::ModuleHotDecline
-      // URL, worker, asset, and CSS resource references do not synchronously
-      // execute the referenced module in the current JavaScript module graph.
-      | DependencyType::NewUrl
-      | DependencyType::NewUrlContext
-      | DependencyType::NewWorker
-      | DependencyType::CreateScriptUrl
-      | DependencyType::CssUrl
-      | DependencyType::CssImport
-      | DependencyType::CssCompose
-      | DependencyType::CssExport
-      | DependencyType::CssLocalIdent
-      | DependencyType::CssSelfReferenceLocalIdent
-      | DependencyType::ExtractCSS
-      // Build-time or metadata-only dependencies.
-      | DependencyType::ExportInfoApi
-      | DependencyType::StaticExports
-      | DependencyType::LoaderImport
-      | DependencyType::ModuleDecorator
-      | DependencyType::RstestModulePath
-      | DependencyType::RstestMockModuleId
-      | DependencyType::RstestHoistMock
-      | DependencyType::RstestDynamicImportOrigin
-  )
-}
-
-#[plugin_hook(CompilationOptimizeModules for CircularModulesInfoPlugin)]
-async fn optimize_modules(
-  &self,
-  compilation: &Compilation,
-  circular_modules: &mut IdentifierSet,
-  _diagnostics: &mut Vec<Diagnostic>,
-) -> Result<Option<bool>> {
-  let module_graph = compilation.get_module_graph();
-  let graph = CycleGraph::build(&module_graph);
-  *circular_modules = CycleDetector::new(&graph).find_circular_modules();
-
-  Ok(None)
-}
-
-impl Plugin for CircularModulesInfoPlugin {
-  fn apply(&self, ctx: &mut rspack_core::ApplyContext<'_>) -> Result<()> {
-    ctx
-      .compilation_hooks
-      .optimize_modules
-      .tap(optimize_modules::new(self));
-    Ok(())
-  }
-}
+pub use circular_modules_info_plugin::CircularModulesInfoPlugin;
